thêm menu sương sương :v, sửa load file, GameController...

+++ src/mrmathami/thegame/Config.java
@@ -24,12 +24,12 @@ public final class Config {
 	 * Num of tiles the screen can display if fieldZoom is TILE_SIZE,
 	 * in other words, the texture will be display as it without scaling.
 	 */
-	public static final long TILE_HORIZONTAL = 30;
+	public static final int TILE_HORIZONTAL = 30;
 	/**
 	 * Num of tiles the screen can display if fieldZoom is TILE_SIZE,
 	 * in other words, the texture will be display as it without scaling.
 	 */
-	public static final long TILE_VERTICAL = 20;
+	public static final int TILE_VERTICAL = 20;
 	/**
 	 * An arbitrary number just to make some code run a little faster.
 	 * Do not touch.
